# Conflicts:
#	apparelrow/apparel/factories.py
#	apparelrow/apparel/tests.py
#	apparelrow/apparel/views/__init__.py

+++ apparelrow/apparel/management/commands/cleanup_vendorcategory.py
@@ -1,9 +1,8 @@
 from optparse import make_option
 
 from django.core.management.base import BaseCommand
-
 from apparelrow.apparel.models import VendorCategory
-
+from progressbar import ProgressBar, Percentage, Bar
 
 class Command(BaseCommand):
     args = ''
@@ -15,15 +14,38 @@ class Command(BaseCommand):
             default=False,
             help='Delete unmapped vendor categories',
         ),
+        make_option('--clean',
+            action='store_true',
+            dest='clean',
+            default=False,
+            help='Cleans out the Vendor Categories from the database.',
+        ),
+        make_option('--verbose',
+            action='store_true',
+            dest='verbose',
+            default=False,
+            help='Shows a progress bar.',
+        ),
     )
 
     def handle(self, *args, **options):
         deleted_count = 0
-        for vc in  VendorCategory.objects.filter(category__isnull=True,
-                                                 override_gender__isnull=True,
-                                                 default_gender__isnull=True).iterator():
-            if options['delete_unmapped'] or vc.vendor_products.count() == 0:
-                vc.delete()
-                deleted_count += 1
+        vcs = VendorCategory.objects.filter(category__isnull=True,override_gender__isnull=True,default_gender__isnull=True)
+        pbar = None
+        if not vcs.count() > 0:
+            print "No vendor categories to clean out."
+            return
 
-        print 'Deleted {0} vendor categories'.format(deleted_count)
+        if options["verbose"]:
+            pbar = ProgressBar(widgets=[Percentage(), Bar()], maxval=vcs.count()).start()
+
+        for index, vc in enumerate(vcs.iterator()):
+            if pbar:
+                pbar.update(index)
+            if options['delete_unmapped'] or vc.vendor_products.count() == 0:
+                if options['clean']:
+                    vc.delete()
+                deleted_count += 1
+        if pbar:
+            pbar.finish()
+        print 'Deleted [{}] {} vendor categories'.format(options['clean'], deleted_count)
